Merge with README.md
Since there's only one file in this folder, it makes sense to merge it with the README. That way the reader doesn't have to click a second time to access the content. This clears up some redundancy as well. if you agree with this suggestion, accept the pull request and merge it. 

This pull request just deletes the duplicate.